requirements.txt
Flask
tensorflow==2.10.0
numpy
Pillow